requirements.txt
bleak==1.1.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
meshcore==2.1.7
meshcore-cli==1.1.23
prompt_toolkit==3.0.52
pycayennelpp==2.4.0
pyserial==3.5
pyserial-asyncio==0.6
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
winrt-runtime==3.2.1
winrt-Windows.Devices.Bluetooth==3.2.1
winrt-Windows.Devices.Bluetooth.Advertisement==3.2.1
winrt-Windows.Devices.Bluetooth.GenericAttributeProfile==3.2.1
winrt-Windows.Devices.Enumeration==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Foundation.Collections==3.2.1
winrt-Windows.Storage.Streams==3.2.1
